A few different commits to improve allocations when running remote write.
1. Remove an unneeded temp variable
1. Only allocate `pendingSamples` once per shard
1. Allocate the `send` slice far fewer times
1. Use a mask rather than always copy samples into a temporary slice
1. Allocate a snappy buffer per shard rather than create one per request.

Added benchmark:
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkSampleDelivery-4     1206794       1141762       -5.39%

benchmark                     old allocs     new allocs     delta
BenchmarkSampleDelivery-4     7152           7121           -0.43%

benchmark                     old bytes     new bytes     delta
BenchmarkSampleDelivery-4     909406        610669        -32.85%
```
Much of the remaining allocs are from proto encoding/decoding.

Reviewing commit by commit will have all changes in isolation.

@cstyan Have you been running any benchmarks for these? I only see one that requires a real WAL somewhere.