@csmarchbanks 2019-05-31
When looking further through the code I am pretty convinced that `seriesMtx` is unnecessary which allows me simplify `Append` to only have one loop and no `tempSamples`. I have been running remote write locally for awhile now with `-race`, and only found the race condition in `maxGauge` that I fixed with the most recent commit.