@csmarchbanks 2019-06-03
Update: I ran this branch all weekend in a cluster without issue. I am reasonably sure that the mutex is not needed.