@tomwilkie 2019-06-04
> Update: I ran this branch all weekend in a cluster without issue. I am reasonably sure that the mutex is not needed.

I think you're right, as there is a single go routine in the watcher driving all these interactions.  This will become more obvious when we refactor the WAL Watcher.

LGTM from me but I'll hold off on merging until @cstyan has had a chance to run this in our env.