@csmarchbanks 2019-06-11
@tomwilkie I know @cstyan ran this for a little bit in one of your environments. Do you need to run it longer or is this good to go?