@cstyan 2019-06-27
I just deployed a new build of this image to some test prometheus instances, it will run overnight at the least and I'll check resource usage in the morning.