@tomwilkie 2019-06-27
LGTM 
<img width="1261" alt="Screenshot 2019-06-27 at 19 47 25" src="https://user-images.githubusercontent.com/444037/60292158-6b4b3f80-9914-11e9-8ab7-063fe7c98c71.png">